# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Notify API Swagger Documentation",
        "title": "Notify API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "0.1"
    },
    "host": "localhost:9999",
    "basePath": "/api/v1/",
    "paths": {
...
        "/notification": {
            "post": {
                "description": "Creates a new notification based on the provided notification data.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Create a new notification",
                "parameters": [
                    {
                        "description": "Notification request body",
                        "name": "notification",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/value.NotificationInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification sent successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable entity",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/notification/{id}": {
            "get": {
                "description": "Retrieves a notification by its unique identifier.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Get notification by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification retrieved successfully",
                        "schema": {
                            "$ref": "#/definitions/value.NotificationInput"
                        }
                    },
                    "404": {
                        "description": "Notification not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "value.Event": {
            "type": "object",
            "required": [
                "category",
                "cost_cents",
                "currency",
                "name",
                "receiver",
                "requester",
                "timestamp"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "cost_cents": {
                    "type": "integer"
                },
                "currency": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "receiver": {
                    "type": "string"
                },
                "requester": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "value.NotificationInput": {
            "type": "object",
            "required": [
                "channels",
                "event",
                "title",
                "uuid"
            ],
            "properties": {
                "channels": {
                    "type": "array",
                    "maxItems": 20,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "event": {
                    "$ref": "#/definitions/value.Event"
                },
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    }
}